requirements.txt
python-telegram-bot==20.7
groq
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
python-dotenv==1.0.1
